Implement meshdb-connected database client
