+++ mesh-database-client/mesh_database_client/__init__.py
@@ -1,1 +1,3 @@
 from .database import DatabaseClient
+from .meshdb_client import MeshDBDatabaseClient
+from .spreadsheet_client import SpreadsheetDatabaseClient
